refactor: update current month store tests for clarity and accuracy
- Removed deprecated `availableToSpend` method references in favor of clearer calculations using `totalAvailableWithRollover` and `totalSpentWithoutRollover`.
- Enhanced test scenarios to better reflect business logic and ensure accurate financial calculations.
- Improved overall readability and maintainability of the test suite.

These changes streamline the testing process and ensure that the current month store accurately represents user financial scenarios.
